+++ cmd/hidden_lake/adapters/common/service/main.go
@@ -29,7 +29,6 @@ func initDB() database.IKVDatabase {
 	db, err = database.NewKeyValueDB(
 		storage.NewSettings(&storage.SSettings{
 			FPath:     databasePath,
-			FHashing:  false,
 			FWorkSize: 1,
 			FPassword: "_",
 		}),