+++ cmd/hidden_lake/messenger/internal/database/keys.go
@@ -6,7 +6,6 @@ import (
 
 const (
 	cKeySizeTemplate          = "database[%s].friends[%s].size"
-	cKeyMessageByUIDTemplate  = "database[%s].friends[%s].messages[uid=%x]"
 	cKeyMessageByEnumTemplate = "database[%s].friends[%s].messages[enum=%d]"
 )
 
@@ -24,14 +23,5 @@ func getKeyMessageByEnum(pR IRelation, pI uint64) []byte {
 		pR.IAm().GetAddress().ToString(),
 		pR.Friend().GetAddress().ToString(),
 		pI,
-	))
-}
-
-func getKeyMessageByUID(pR IRelation, pBlockUID [cBlockUIDSize]byte) []byte {
-	return []byte(fmt.Sprintf(
-		cKeyMessageByUIDTemplate,
-		pR.IAm().GetAddress().ToString(),
-		pR.Friend().GetAddress().ToString(),
-		pBlockUID[:],
 	))
 }