+++ cmd/hidden_lake/messenger/internal/database/message.go
@@ -4,14 +4,12 @@ import (
 	"bytes"
 	"time"
 
-	"github.com/number571/go-peer/pkg/crypto/random"
 	"github.com/number571/go-peer/pkg/encoding"
 )
 
 const (
 	cIsIncomingSize = 1
 	cTimestampSize  = encoding.CSizeUint64
-	cBlockUIDSize   = 16
 )
 
 var (
@@ -21,24 +19,19 @@ var (
 type sMessage struct {
 	fIsIncoming bool
 	fTimestamp  uint64
-	fBlockUID   [cBlockUIDSize]byte
 	fMessage    []byte
 }
 
 func NewMessage(pIsIncoming bool, pMessage []byte) IMessage {
-	blockUID := [cBlockUIDSize]byte{}
-	copy(blockUID[:], random.NewStdPRNG().GetBytes(cBlockUIDSize))
-
 	return &sMessage{
 		fIsIncoming: pIsIncoming,
-		fBlockUID:   blockUID,
 		fTimestamp:  uint64(time.Now().Unix()),
 		fMessage:    pMessage,
 	}
 }
 
 func LoadMessage(pMsgBytes []byte) IMessage {
-	if len(pMsgBytes) < (cIsIncomingSize + cTimestampSize + cBlockUIDSize) {
+	if len(pMsgBytes) < (cIsIncomingSize + cTimestampSize) {
 		return nil
 	}
 
@@ -50,23 +43,15 @@ func LoadMessage(pMsgBytes []byte) IMessage {
 	blockTimestamp := [cTimestampSize]byte{}
 	copy(blockTimestamp[:], pMsgBytes[cIsIncomingSize:cIsIncomingSize+cTimestampSize])
 
-	blockUID := [cBlockUIDSize]byte{}
-	copy(blockUID[:], pMsgBytes[cIsIncomingSize:cIsIncomingSize+cTimestampSize:cIsIncomingSize+cTimestampSize+cBlockUIDSize])
-
 	return &sMessage{
 		fIsIncoming: isIncoming,
-		fBlockUID:   blockUID,
 		fTimestamp:  encoding.BytesToUint64(blockTimestamp),
-		fMessage:    pMsgBytes[cIsIncomingSize+cTimestampSize+cBlockUIDSize:],
+		fMessage:    pMsgBytes[cIsIncomingSize+cTimestampSize:],
 	}
 }
 
 func (p *sMessage) IsIncoming() bool {
 	return p.fIsIncoming
-}
-
-func (p *sMessage) GetBlockUID() [cBlockUIDSize]byte {
-	return p.fBlockUID
 }
 
 func (p *sMessage) GetMessage() []byte {
@@ -87,7 +72,6 @@ func (p *sMessage) ToBytes() []byte {
 		[][]byte{
 			{firstByte},
 			blockTimestamp[:],
-			[]byte(p.fBlockUID[:]),
 			[]byte(p.fMessage),
 		},
 		[]byte{},