+++ cmd/hidden_lake/messenger/pkg/settings/settings.go
@@ -17,8 +17,9 @@ const (
 )
 
 const (
-	CIamAliasName    = "__iam__"
-	CDefaultLanguage = "ENG"
+	CIamAliasName      = "__iam__"
+	CDefaultLanguage   = "ENG"
+	CDefaultStorageKey = "_"
 )
 
 const (